relocated updatelinksvalue api to modifylinks path
